All Tests are passing. FAQs are posting in Test mode.

+++ src/main/java/com/revature/movieapp/repos/FAQs_Repo.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FAQs_Repo extends JpaRepository<FAQs,Long> {
-    //This is also referred to as a Data-Access-Object.
+public interface FAQs_Repo extends JpaRepository<FAQs,Integer> {
+
            }
